更新jackson

+++ src/main/java/org/springframework/social/weibo/connect/WeiboOAuth2Template.java
@@ -15,16 +15,10 @@
  */
 package org.springframework.social.weibo.connect;
 
-import java.io.IOException;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Map.Entry;
-
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.codehaus.jackson.map.ObjectMapper;
-import org.codehaus.jackson.type.TypeReference;
 import org.springframework.http.HttpInputMessage;
 import org.springframework.http.MediaType;
 import org.springframework.http.converter.FormHttpMessageConverter;
@@ -37,67 +31,73 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 
+import java.io.IOException;
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.Map;
+import java.util.Map.Entry;
+
 public class WeiboOAuth2Template extends OAuth2Template {
 
-	private static final Log logger = LogFactory
-			.getLog(WeiboOAuth2Template.class.getName());
+    private static final Log logger = LogFactory
+            .getLog(WeiboOAuth2Template.class.getName());
 
-	public WeiboOAuth2Template(String clientId, String clientSecret) {
-		super(clientId, clientSecret,
-				"https://api.t.sina.com.cn/oauth2/authorize",
-				"https://api.t.sina.com.cn/oauth2/access_token");
-	}
+    public WeiboOAuth2Template(String clientId, String clientSecret) {
+        super(clientId, clientSecret,
+                "https://api.t.sina.com.cn/oauth2/authorize",
+                "https://api.t.sina.com.cn/oauth2/access_token");
+    }
 
-	@Override
-	protected RestTemplate createRestTemplate() {
-		RestTemplate restTemplate = new RestTemplate(
-				ClientHttpRequestFactorySelector.getRequestFactory());
-		HttpMessageConverter<?> messageConverter = new FormHttpMessageConverter() {
+    @Override
+    protected RestTemplate createRestTemplate() {
+        RestTemplate restTemplate = new RestTemplate(
+                ClientHttpRequestFactorySelector.getRequestFactory());
+        HttpMessageConverter<?> messageConverter = new FormHttpMessageConverter() {
 
-			private final ObjectMapper objectMapper = new ObjectMapper();
+            private final ObjectMapper objectMapper = new ObjectMapper();
 
-			@Override
-			public boolean canRead(Class<?> clazz, MediaType mediaType) {
-				return true;
-			}
+            @Override
+            public boolean canRead(Class<?> clazz, MediaType mediaType) {
+                return true;
+            }
 
-			@Override
-			public MultiValueMap<String, String> read(
-					Class<? extends MultiValueMap<String, ?>> clazz,
-					HttpInputMessage inputMessage) throws IOException,
-					HttpMessageNotReadableException {
+            @Override
+            public MultiValueMap<String, String> read(
+                    Class<? extends MultiValueMap<String, ?>> clazz,
+                    HttpInputMessage inputMessage) throws IOException,
+                    HttpMessageNotReadableException {
 
-				TypeReference<Map<String, ?>> mapType = new TypeReference<Map<String, ?>>() {
-				};
-				LinkedHashMap<String, ?> readValue = objectMapper.readValue(
-						inputMessage.getBody(), mapType);
-				LinkedMultiValueMap<String, String> result = new LinkedMultiValueMap<String, String>();
-				for (Entry<String, ?> currentEntry : readValue.entrySet()) {
-					result.add(currentEntry.getKey(), currentEntry.getValue()
-							.toString());
-				}
-				return result;
-			}
-		};
+                TypeReference<Map<String, ?>> mapType = new TypeReference<Map<String, ?>>() {
+                };
+                LinkedHashMap<String, ?> readValue = objectMapper.readValue(
+                        inputMessage.getBody(), mapType);
+                LinkedMultiValueMap<String, String> result = new LinkedMultiValueMap<String, String>();
+                for (Entry<String, ?> currentEntry : readValue.entrySet()) {
+                    result.add(currentEntry.getKey(), currentEntry.getValue()
+                            .toString());
+                }
+                return result;
+            }
+        };
 
-		restTemplate.setMessageConverters(Collections
-				.<HttpMessageConverter<?>> singletonList(messageConverter));
-		return restTemplate;
-	}
+        restTemplate.setMessageConverters(Collections
+                .<HttpMessageConverter<?>>singletonList(messageConverter));
+        return restTemplate;
+    }
 
-	@Override
-	@SuppressWarnings("unchecked")
-	protected AccessGrant postForAccessGrant(String accessTokenUrl,
-			MultiValueMap<String, String> parameters) {
-		MultiValueMap<String, String> response = getRestTemplate()
-				.postForObject(accessTokenUrl, parameters, MultiValueMap.class);
-		String expires = response.getFirst("expires_in");
-		String accessToken = response.getFirst("access_token");
-		if (logger.isDebugEnabled()) {
-			logger.debug("access token value = " + accessToken);
-		}
-		return new AccessGrant(accessToken, null, null,
-				expires != null ? Integer.valueOf(expires) : null);
-	}
+    @Override
+    @SuppressWarnings("unchecked")
+    protected AccessGrant postForAccessGrant(String accessTokenUrl,
+                                             MultiValueMap<String, String> parameters) {
+        MultiValueMap<String, String> response = getRestTemplate()
+                .postForObject(accessTokenUrl, parameters, MultiValueMap.class);
+        String expires = response.getFirst("expires_in");
+        String accessToken = response.getFirst("access_token");
+        if (logger.isDebugEnabled()) {
+            logger.debug("access token value = " + accessToken);
+        }
+        return new AccessGrant(accessToken, null, null,
+                expires != null ? Long.valueOf(expires) : null);
+    }
 
 }
